week1/python-full-code/challenge-code/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0 